ultimas versiones a entregr de libft

+++ 0-Libft/ENTREGA/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: nazuaje- <nazuaje-@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/11 15:30:10 by nazuaje-          #+#    #+#             */
-/*   Updated: 2025/10/24 18:54:13 by nazuaje-         ###   ########.fr       */
+/*   Updated: 2025/10/26 13:49:17 by nazuaje-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 static int	ft_get_number(const char *str, int sign)
 {
-	long	result;
+	int	result;
 	char	*s;
 
 	result = 0;
@@ -24,7 +24,7 @@ static int	ft_get_number(const char *str, int sign)
 		result = (result * 10) + *s - '0';
 		s++;
 	}
-	return ((int)result * sign);
+	return (result * sign);
 }
 
 int	ft_atoi(const char *str)
